+++ test/multi-map.test.ts
@@ -1,7 +1,4 @@
 import { MultiLevelMap } from '../src/util/multi-level-map';
-import { Utility } from '../src/util/utility';
-
-const util = new Utility();
 
 describe('multi-level map use cases', () => {
       
@@ -37,12 +34,5 @@ describe('multi-level map use cases', () => {
         expect(mm.getElement('a.b.c[4].test2')).toBe('message2');    
         expect(mm.exists('test.boolean')).toBe(true);
         expect(mm.getElement('test.boolean')).toBe(false);
-    }); 
-    
-    it('can detect path traversal', () => {
-        expect(() => {
-            util.getSafeFilePath('/tmp', '../hello');
-        }).toThrow('Access denied because file path is outside the base directory'); 
-    });  
-
+    });
 });